primary-requirements.txt
BeautifulSoup
CleverCSS
Django
Jinja2
Markdown
PyYAML
Pygments
South
Werkzeug
amqp
anyjson
argparse
billiard
celery
chardet
django-celery
django-debug-toolbar
django-extensions
django-filter
djangorestframework
flake8
httpie
ipython
kombu
lmtpd
lockfile
mccabe
mock
nose
pep8
pika
pydns
pyflakes
python-daemon
python-dateutil
python-modargs
pytz
requests
salmon-mail
six
sqlparse
wsgiref
